--- a/Foundation/NSJSONSerialization.swift
+++ b/Foundation/NSJSONSerialization.swift
@@ -420,7 +420,23 @@ private struct JSONDeserializer {
     }
     
     //MARK: - Number parsing
-    static let numberScalars = ".+-0123456789eE".unicodeScalars
+    static let numberScalars = [
+        UnicodeScalar(0x2E), // .
+        UnicodeScalar(0x30), // 0
+        UnicodeScalar(0x31), // 1
+        UnicodeScalar(0x32), // 2
+        UnicodeScalar(0x33), // 3
+        UnicodeScalar(0x34), // 4
+        UnicodeScalar(0x35), // 5
+        UnicodeScalar(0x36), // 6
+        UnicodeScalar(0x37), // 7
+        UnicodeScalar(0x38), // 8
+        UnicodeScalar(0x39), // 9
+        UnicodeScalar(0x65), // e
+        UnicodeScalar(0x45), // E
+        UnicodeScalar(0x2B), // +
+        UnicodeScalar(0x2D), // -
+    ]
     static func parseNumber(input: UnicodeParser) throws -> (Double, UnicodeParser)? {
         let view = input.view
         let endIndex = view.endIndex