I ran 2 benchmarks against the current JSON parser implementation on `master`, comparing it to Darwin Foundation for [large inputs](https://github.com/zemirco/sf-city-lots-json/blob/master/citylots.json) and found it was 5-10x slower. (72s vs 8s for a single decode). This new code reduces this to a factor for 2-4x (17s)

While JSON is a text-based serialization, all control characters are ASCII characters. The data can be most efficiently parsed as a stream of bytes instead of the initial overhead of converting the bytes to a String.

The byte -> string conversion in the current code only accounts for 3s of time but parsing numbers is incredibly expensive (`Double.init?(_ text: String)` converts the string back to bytes to use [`strtod()`](https://developer.apple.com/library/mac/documentation/Darwin/Reference/ManPages/man3/strtod.3.html#//apple_ref/doc/man/3/strtod)). This technique uses `strtod` directly on the byte array. There is a penalty for non-UTF-8 encoded data which is inline with the Darwin implementation

> The most efficient encoding to use for parsing is UTF-8, so if you have a choice in encoding the data passed to this method, use UTF-8.

Another modification has been made to remove the intermediate `parser` values as the heap allocations for each intermediate parser was adding significant overhead.

This also includes a commit to implement `.AllowFragments`.