@phausler 2015-12-29
from a cursory glance this looks like a reasonable change; builds pass on ubuntu.

Do you happen to have the performance benchmarks handy? that might be something interesting to check in so that we could potentially build up a performance test suite.

Additionally what other hotspots are there in Foundation that you found by this?