@argon 2015-12-29
The performance "benchmark" wasn't in any way automated unfortunately. I simply created a `main.swift` to read the JSON file from disk then decode it, then I profiled it using instruments.

The main hotspots in _this_ code now relate to

 - Array and Dictionary size increases, leading to a large number of copies
 - Bridging from NSString -> String in `parseString`

In general String parsing for the sf-city-lots-json bears a huge time cost ~(5000ms vs 250ms) when compared with Darwin Foundation.